Move out and rewrite UseTree completion tests
